+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  * Add --binary support to 'fx mag'
  * Use unit-free prefixes by default, "b" for units with --binary
  * Remove run() from range.rs
+ * Replace rename() and xargs() stubs in main.rs with modules
+
 0.0.4 ... 2018.0120 13:08:21
 
  * Update 'fx range help' output to be consistent